storage: Fix hadoop tests.

+++ opencga-storage/opencga-storage-hadoop/opencga-storage-hadoop-core/src/test/java/org/opencb/opencga/storage/hadoop/variant/VariantHbaseTestUtils.java
@@ -401,6 +401,10 @@ public class VariantHbaseTestUtils {
     private static void printSamplesIndexTable(VariantHadoopDBAdaptor dbAdaptor, Path outDir) throws IOException {
         for (Integer studyId : dbAdaptor.getStudyConfigurationManager().getStudies(null).values()) {
             String sampleGtTableName = dbAdaptor.getTableNameGenerator().getSampleIndexTableName(studyId);
+            if (!dbAdaptor.getHBaseManager().tableExists(sampleGtTableName)) {
+                // Skip table
+                return;
+            }
             Path fileName = outDir.resolve(sampleGtTableName + ".txt");
             try (
                     FileOutputStream fos = new FileOutputStream(fileName.toFile()); PrintStream out = new PrintStream(fos)
